+++ lib/pages/other/OrganisationPage.dart
@@ -1,9 +1,12 @@
-import 'package:app/assets/StyleFrom.dart';
 import 'package:app/assets/components/header.dart';
 import 'package:app/assets/components/sectionTitle.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+
+import 'package:app/assets/components/darkButton.dart';
+import 'package:app/assets/components/campaignTile.dart';
+import 'package:app/assets/routes/customLaunch.dart';
 
 import 'package:app/models/Organisation.dart';
 
@@ -56,6 +59,32 @@ class OraganisationInfoPage extends StatelessWidget {
                 Text(
                   organisation.getDescription(),
                 ),
+              
+                organisation.getCampaigns().length > 0 
+                  ? Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        SizedBox(height: BETWEEN_SECTION_PADDING),
+                        SectionTitle(
+                          "Associated campaigns",
+                          vpadding: SECTION_TITLE_BOTTOM_PADDING,
+                        ),
+                        Container(
+                          height: 120,
+                          child: PageView.builder(
+                            itemCount: organisation.getCampaigns().length,
+                            itemBuilder: (BuildContext context, int index) {
+                              return CampaignSelectionTile(
+                                organisation.getCampaigns()[index],
+                                height: 120,
+                              );
+                            },
+                          ),
+                        ),
+                      ]
+                    )
+                  : Container()
+                ,
                 
                 SizedBox(height: BETWEEN_SECTION_PADDING),
                
@@ -109,9 +138,9 @@ class OraganisationInfoPage extends StatelessWidget {
                 SectionTitle(
                   "Find out more",
                 ),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceAround,
-                  mainAxisSize: MainAxisSize.max,
+                Column(
+                  mainAxisSize: MainAxisSize.min,
+                  crossAxisAlignment: CrossAxisAlignment.center,
                   children: getExtraLinks(organisation),
                 ),
 
@@ -171,13 +200,13 @@ class SocialMediaButton extends StatelessWidget {
 List<Widget> getExtraLinks(Organisation org) {
   List<Widget> extraLinks = [];
   if (org.getExtraText1() != null) {
-    extraLinks.add(ExtraLinkButton(org.getExtraText1(), org.getExtraLink1(), FontAwesomeIcons.externalLinkAlt));
+    extraLinks.add(ExtraLinkButton(org.getExtraText1(), org.getExtraLink1()));
   }
   if (org.getExtraText2() != null) {
-    extraLinks.add(ExtraLinkButton(org.getExtraText2(), org.getExtraLink2(), FontAwesomeIcons.externalLinkAlt));
+    extraLinks.add(ExtraLinkButton(org.getExtraText2(), org.getExtraLink2()));
   }
   if (org.getExtraText3() != null) {
-    extraLinks.add(ExtraLinkButton(org.getExtraText3(), org.getExtraLink3(), FontAwesomeIcons.externalLinkAlt));
+    extraLinks.add(ExtraLinkButton(org.getExtraText3(), org.getExtraLink3()));
   }
   return extraLinks;
 }
@@ -185,47 +214,16 @@ List<Widget> getExtraLinks(Organisation org) {
 class ExtraLinkButton extends StatelessWidget {
   final String text;
   final String link;
-  final IconData icon;
-  ExtraLinkButton(this.text, this.link, this.icon);
+  ExtraLinkButton(this.text, this.link);
   @override
   Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: () {
-        launch(link);
-      },
-      child: Container(
-        width: MediaQuery.of(context).size.width * 0.3,
-        decoration: BoxDecoration(
-          color: Colors.white,
-          boxShadow: [
-            BoxShadow(
-              color: Color.fromRGBO(0, 45, 96, 0.08),
-              offset: Offset(0, 3),
-              blurRadius: 20,
-            ),
-          ],
-        ),
-        child: Padding(
-          padding: EdgeInsets.all(10),
-          child: Column(
-            children: [
-              Icon(
-                icon,
-                color: Color.fromRGBO(155,159,177,1),
-                size: 25,
-              ),
-              SizedBox(height: 10),
-              Text(
-                text,
-                style: textStyleFrom(
-                  Theme.of(context).primaryTextTheme.bodyText1,
-                  color: Color.fromRGBO(155,159,177,1),
-                ),
-              )
-            ],
-          ),
-        )
-    )
-  );
+    return Padding(
+      padding: EdgeInsets.all(8),
+      child: DarkButton(
+        text,
+        onPressed: () {customLaunch(context, link);},
+        inverted: true,
+      ),
+    );
   }
 }
